Change 'RoleId' property data type from byte to int + Update 'Role' and 'User' models and DbContext

+++ api/SmartMeetingRoom.API/Models/Role.cs
@@ -5,7 +5,7 @@ namespace SmartMeetingRoom.API.Models;
 
 public partial class Role
 {
-    public byte RoleId { get; set; }
+    public int RoleId { get; set; }
 
     public string RoleName { get; set; } = null!;
